+++ tests/utils/test_discovery.py
@@ -30,6 +30,12 @@ anomaly_detectors = [
     anomaly_detection.ConvolutionalNeuralNetwork,
     anomaly_detection.LongShortTermMemoryNetwork,
     anomaly_detection.Transformer,
+    anomaly_detection.LocalPolynomialApproximation,
+    anomaly_detection.ChronosAnomalyDetector,
+    anomaly_detection.SpectralResidual,
+    anomaly_detection.MOMENTAnomalyDetector,
+    anomaly_detection.TimeMoEAnomalyDetector,
+    anomaly_detection.RobustRandomCutForestAnomalyDetector,
 ]
 data_loaders = [
     data.DemonstrationTimeSeriesLoader,
@@ -59,6 +65,7 @@ metrics = [
     evaluation.RangeBasedPrecision,
     evaluation.RangeBasedRecall,
     evaluation.RangeBasedFBeta,
+    evaluation.UCRScore
 ]
 proba_metrics = [
     evaluation.ThresholdMetric,
@@ -69,6 +76,7 @@ proba_metrics = [
     evaluation.RangeAreaUnderPR,
     evaluation.VolumeUnderROC,
     evaluation.VolumeUnderPR,
+    evaluation.UCRScore
 ]
 binary_metrics = [
     evaluation.Precision,